Calculate holidays in a specific timezone
Holidays unlike an appointment are based on days. So they care
if the 1.5.2012 is a holiday or not. Therefore a holiday is specified
by a date and not by it's timezone. To calculate holidays for a given
timezone we have to pass the timezone parameter. This will result
in the all holidays starting at 0:00 for the specified timezone.

E.g. if you calculate Germany holidays on 1.5. on PST timezone
you will get a Datetime "Tag der Arbeit" starting at 0:00 PST time.

If you pass "CET" you will get "Tag der Arbeit" starting at 0:00 CET.