Test against recent PHP 7.x versions; refactor PHP version constraint
Adds the most recent PHP versions of PHP to ensure compatibility with
supported versions of PHP.

Additionally the versions constraint for PHP itself has been refactored,
it's generally not a good idea to specify unbound constraints that
blindly support all future releases of a dependency (or PHP in this case).
Instead the constraint should be carefully modified after new releases
when compliance is ensured.